fix: PR, typo
